feat: add ABIEncode

+++ contracts/ABIEncode.sol
@@ -2,25 +2,27 @@
 pragma solidity ^0.8.23;
 
 interface IERC20 {
-    function transfer(address, uint) external ;
+    function transfer(address, uint256) external ;
 }
 
 contract Token {
-    function transfer(address, uint) external  {} 
+    function transfer(address, uint256) external  {} 
 }
 contract AbiEncode {
-    // Test  
+   // Test  
    function test(address _contract, bytes calldata data) external {
       (bool ok, ) = _contract.call(data);
       require(ok, "call failed");
    }
+
    // encode with singnature
-   function encodeWithSingature(address to , uint amount) external pure returns (bytes memory) {
+   function encodeWithSingature(address to , uint256 amount) external pure returns (bytes memory) {
       // encodeWithSignature adddress amount
       return abi.encodeWithSignature("transfer(address, uint256)", to , amount);
    }
+
    // encdoe with selector
-   function encodeWithSelector(address to, uint amount) 
+   function encodeWithSelector(address to, uint256 amount) 
    external  
    pure 
    returns (bytes memory) {
@@ -29,7 +31,7 @@ contract AbiEncode {
      return abi.encodeWithSelector(IERC20.transfer.selector, to, amount);
    }
     // encode call 
-    function encodeCall(address to, uint amount) external  pure returns (bytes memory) {
+    function encodeCall(address to, uint256 amount) external  pure returns (bytes memory) {
         // Typo and type errors will not compile
         return  abi.encodeCall(IERC20.transfer, (to, amount));
     }
